code/requirements.txt
asttokens==2.4.1
bio==1.6.0
biopython==1.82
biothings-client==0.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
ctranslate2==3.23.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fasttext==0.9.2
gprofiler-official==1.0.0
idna==3.6
invoke==2.2.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
Levenshtein==0.23.0
matplotlib-inline==0.1.6
mygene==3.2.2
mypy==1.8.0
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
pooch==1.8.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.11.1
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
requests==2.31.0
ruff==0.1.11
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
